# bpo-41361: Optimized argument parsing for deque_rotate