--- a/Misc/NEWS.d/next/Library/2021-03-13-08-18-01.bpo-41361.lXDIlr.rst
+++ b/Misc/NEWS.d/next/Library/2021-03-13-08-18-01.bpo-41361.lXDIlr.rst
@@ -1,0 +1,1 @@
+:meth:`~collections.deque.rotate` calls are now slightly faster due to faster argument parsing.