@sweeneyde 2021-03-13
@rhettinger FWIW, with the refleak fixed, I got these benchmarks

```
Before, with PGO:

    rotate()   --> Mean +- std dev: 50.3 ns +- 1.9 ns
    rotate(1)  --> Mean +- std dev: 61.9 ns +- 2.7 ns
    rotate(-1) --> Mean +- std dev: 63.7 ns +- 0.8 ns

After, with PGO:

    rotate()   --> Mean +- std dev: 40.9 ns +- 0.4 ns
    rotate(1)  --> Mean +- std dev: 46.9 ns +- 0.3 ns
    rotate(-1) --> Mean +- std dev: 47.4 ns +- 0.3 ns
```